backend/tests/security/test_authentication.py
"""Authentication Security Tests for Fynlo POS.

Tests authentication flows, token validation, and session security.
"""

from datetime import datetime, timedelta
from unittest.mock import MagicMock, Mock, patch
from uuid import uuid4

import pytest
from fastapi import HTTPException
from jose import jwt

from app.core.auth import get_current_user, get_platform_owner

# Note: The following functions need to be imported from their actual locations
# or mocked for testing purposes:
# - create_access_token (may be in a JWT utility module)
# - verify_token (may be in a JWT utility module)
# These tests assume these functions exist for demonstration purposes
from app.core.config import settings
from app.models.user import User


# Mock implementations for testing
def create_access_token(data: dict, expires_delta: timedelta = None):
    """Mock implementation of create_access_token for testing."""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire, "iat": datetime.utcnow()})
    encoded_jwt = jwt.encode(
        to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM
    )
    return encoded_jwt


def verify_token(token: str):
    """Mock implementation of verify_token for testing."""
    try:
        payload = jwt.decode(
            token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM]
        )
        if "sub" not in payload:
            raise HTTPException(status_code=401, detail="Invalid token")
        return payload
    except jwt.JWTError:
        raise HTTPException(status_code=401, detail="Invalid token")


@pytest.fixture
def mock_user():
    """Create a mock user for testing."""
    user = Mock(spec=User)
    user.id = uuid4()
    user.email = "test@restaurant.com"
    user.role = "restaurant_owner"
    user.restaurant_id = uuid4()
    user.is_active = True
    user.hashed_password = "$2b$12$hashedpassword"
    return user


@pytest.fixture
def mock_platform_owner():
    """Create a mock platform owner for testing."""
    user = Mock(spec=User)
    user.id = uuid4()
    user.email = "platform@fynlo.com"
    user.role = "platform_owner"
    user.restaurant_id = None
    user.is_active = True
    return user


class TestAuthenticationSecurity:
    """Test suite for authentication security."""

    def test_token_creation_includes_required_claims(self, mock_user):
        """Test that tokens include all required security claims."""
        token_data = {
            "sub": str(mock_user.id),
            "email": mock_user.email,
            "role": mock_user.role,
            "restaurant_id": (
                str(mock_user.restaurant_id)
                if mock_user.restaurant_id
                else None
            ),
        }

        token = create_access_token(data=token_data)

        # Decode token without verification to check claims
        decoded = jwt.decode(token, options={"verify_signature": False})

        assert "sub" in decoded
        assert "email" in decoded
        assert "role" in decoded
        assert "exp" in decoded
        assert "iat" in decoded  # Issued at time

    def test_token_expiration_enforced(self):
        """Test that expired tokens are rejected."""
        # Create token that expires immediately
        token_data = {"sub": "test_user", "email": "test@test.com"}
        expires_delta = timedelta(seconds=-1)  # Already expired

        token = create_access_token(
            data=token_data, expires_delta=expires_delta
        )

        with pytest.raises(HTTPException) as exc_info:
            verify_token(token)
        assert exc_info.value.status_code == 401

    def test_token_signature_validation(self):
        """Test that tokens with invalid signatures are rejected."""
        token_data = {"sub": "test_user", "email": "test@test.com"}
        valid_token = create_access_token(data=token_data)

        # Tamper with the token
        parts = valid_token.split(".")
        tampered_token = f"{parts[0]}.tampered_payload.{parts[2]}"

        with pytest.raises(HTTPException) as exc_info:
            verify_token(tampered_token)
        assert exc_info.value.status_code == 401

    def test_malformed_token_rejected(self):
        """Test that malformed tokens are rejected."""
        malformed_tokens = [
            "not.a.token",
            "invalid_format",
            "",
            "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9",  # Missing parts
        ]

        for token in malformed_tokens:
            with pytest.raises(HTTPException) as exc_info:
                verify_token(token)
            assert exc_info.value.status_code == 401

    def test_missing_required_claims_rejected(self):
        """Test that tokens missing required claims are rejected."""
        # Token without 'sub' claim
        incomplete_token = jwt.encode(
            {
                "email": "test@test.com",
                "exp": datetime.utcnow() + timedelta(hours=1),
            },
            settings.SECRET_KEY,
            algorithm=settings.ALGORITHM,
        )

        with pytest.raises(HTTPException) as exc_info:
            verify_token(incomplete_token)
        assert exc_info.value.status_code == 401

    @patch("app.core.auth.get_user_by_id")
    def test_inactive_user_rejected(self, mock_get_user, mock_user):
        """Test that inactive users cannot authenticate."""
        mock_user.is_active = False
        mock_get_user.return_value = mock_user

        token = create_access_token(data={"sub": str(mock_user.id)})

        with pytest.raises(HTTPException) as exc_info:
            get_current_user(token, MagicMock())
        assert exc_info.value.status_code == 401

    @patch("app.core.auth.get_user_by_id")
    def test_deleted_user_rejected(self, mock_get_user):
        """Test that deleted users cannot authenticate."""
        mock_get_user.return_value = None  # User not found

        token = create_access_token(data={"sub": str(uuid4())})

        with pytest.raises(HTTPException) as exc_info:
            get_current_user(token, MagicMock())
        assert exc_info.value.status_code == 401

    def test_platform_owner_validation(self, mock_platform_owner, mock_user):
        """Test that only platform owners can access platform endpoints."""
        # Platform owner should succeed
        result = get_platform_owner(mock_platform_owner)
        assert result == mock_platform_owner

        # Non-platform owner should fail
        with pytest.raises(HTTPException) as exc_info:
            get_platform_owner(mock_user)
        assert exc_info.value.status_code == 403
        assert "Platform owner access required" in exc_info.value.detail

    def test_role_cannot_be_elevated_via_token(self):
        """Test that users cannot elevate their role by manipulating tokens."""
        # Create token for regular user
        token_data = {
            "sub": str(uuid4()),
            "email": "user@restaurant.com",
            "role": "employee",
        }

        token = create_access_token(data=token_data)

        # Try to decode and modify the role
        # This should fail due to signature validation
        decoded = jwt.decode(token, options={"verify_signature": False})
        decoded["role"] = "platform_owner"

        # Create new token with elevated role (without proper key)
        fake_key = "fake_secret_key"
        elevated_token = jwt.encode(
            decoded, fake_key, algorithm=settings.ALGORITHM
        )

        # Should fail verification
        with pytest.raises(HTTPException) as exc_info:
            verify_token(elevated_token)
        assert exc_info.value.status_code == 401

    def test_token_reuse_after_password_change(self):
        """Test that old tokens should be invalidated after password change."""
        # This test demonstrates the need for token blacklisting or versioning
        # Currently, tokens remain valid until expiration even after
        # password change
        # This is a known limitation that should be documented

        # TODO: Implement token blacklisting or user token version
        # to invalidate tokens after password change
        pass

    def test_concurrent_session_limit(self):
        """Test that users cannot have unlimited concurrent sessions."""
        # This test demonstrates the need for session management
        # Currently, there's no limit on concurrent sessions
        # This could be a security concern for shared accounts

        # TODO: Implement session management to limit concurrent logins
        pass

    def test_brute_force_protection(self):
        """Test that authentication endpoints have brute force protection."""
        # This should be implemented at the endpoint level
        # Using rate limiting middleware

        # TODO: Verify rate limiting is applied to auth endpoints
        pass

    @patch("app.core.auth.get_user_by_id")
    def test_cross_restaurant_token_usage(self, mock_get_user, mock_user):
        """Test that tokens cannot be used across different restaurants."""
        original_restaurant = mock_user.restaurant_id

        # User switches restaurant (shouldn't be possible, but test anyway)
        mock_user.restaurant_id = uuid4()
        mock_get_user.return_value = mock_user

        # Token still has original restaurant_id
        token_data = {
            "sub": str(mock_user.id),
            "restaurant_id": str(original_restaurant),
        }
        create_access_token(data=token_data)

        # This should be detected and handled
        # Current implementation may not catch this - documenting for
        # future fix
        # TODO: Implement restaurant_id validation in token verification
        pass

    def test_token_algorithm_restriction(self):
        """Test that only allowed algorithms are accepted."""
        # Create token with different algorithm
        token_data = {
            "sub": "test_user",
            "exp": datetime.utcnow() + timedelta(hours=1),
        }

        # Try to use 'none' algorithm (security vulnerability)
        none_token = jwt.encode(token_data, "", algorithm="none")

        with pytest.raises(HTTPException):
            verify_token(none_token)

    def test_token_audience_validation(self):
        """Test that tokens are validated for correct audience."""
        # If audience claim is used, verify it's checked
        # This prevents tokens from other services being used

        # TODO: Implement audience validation if using multiple services
        pass
